Introduce a global tracking script for the platform
This uses Universal analytics rather than ga.js, so it *should* not
conflict with the current functionality offered by jTrack. From
what I can tell, there is no previous ga object created, so we can
create our own and go for it.

We do need to set the cookie domain to wsu.edu to globally track any
subdomains. We will have issues tracking other root domains because
of the way cookies work. We'll cross that bridge when we get there.