Provide defaults for analytics options and merge with DB results
The default passed to `get_option()` is overwritten if anything is
stored in the database at all. If for some reason an empty value
was stored or different data came back from the call somehow, we
should parse this against a set of defaults to make sure all of
the values are available to us.