Only watch the jQuery UI flag before adding UI events
There is more than one way to load jQuery UI into a theme. In the
Spine parent, we use a custom implementation due to the Spine's
requirements. This should be an option to disable or enable.

In the future, it would be nice to merge these in without an extra
HTTP request. All of this event data *could* be localized with the
tracker data, or we could generate a single config.